[Performance] Transition RetinaEncoder to integer math pipeline for FPU-free vision processing * Implement uint8/int16 buffers in RetinaEncoder to eliminate FPU overhead and L2 cache thrashing * Replace float32 Gaussian buffers with int16 for _center, _surround, _dog, _motion, and chromatic opponent channels * Convert base_threshold to integer and use int() for dynamic threshold calculation * Replace float multiplication with np.right_shift for fast divide-by-2 in B-Y opponent computation * Extract BGR channels directly into int16 buffers via uint8 frame copy instead of cv2.extractChannel
[Performance] Upgrade all encoder/decoder pipelines from float16 to float32 precision
* Transition PwmEncoder, PwmDecoder, PopulationEncoder, PopulationDecoder to float32 dtype for all buffers and operations
* Update benchmark_encoders.py, test_population_decoder.py, test_zero_gc.py to use float32 input data
* Change np.linspace centers and phase arrays in encoders from float16 to float32

[Architecture] Align warp size to 64 threads for AMD Wavefront and L2 cache line compatibility
* Update BrainBuilder warp alignment from 32 to 64 threads for padded_n and total_axons calculations
* Adjust RetinaEncoder C-ABI warp alignment from 32-bit to 64-bit multiples for AMD/L2 compatibility

[Performance] Optimize Zero-Copy padding with static buffer allocation in axic-baker
* Replace dynamic vec allocation with static [u8; 4096] ZEROS array for file alignment padding in pack_directory_to_axic()

[Fix] Add hardware VRAM guard in AMD physics kernel
* Insert target_id >= total_axons bounds check in cu_apply_gsop_kernel to prevent out-of-bounds access

Tested: benchmark_encoders.py, test_population_decoder.py, test_zero_gc.py
Refs: RetinaEncoder, PwmEncoder, PwmDecoder, PopulationEncoder, PopulationDecoder, BrainBuilder, axic-baker, physics.hip

+++ axicor-client/axicor/decoders.py
@@ -12,11 +12,11 @@ class PwmDecoder:
         
         # Payload size: B ticks * N motors (1 byte = 1 spike flag)
         self.payload_size = self.N * self.B
-        self._inv_b = np.float16(1.0 / self.B)
+        self._inv_b = np.float32(1.0 / self.B)
         
         # Preallocation for HFT cycle (Zero-Garbage)
-        self._sum_buffer = np.zeros(self.N, dtype=np.float16)
-        self._out_buffer = np.zeros(self.N, dtype=np.float16)
+        self._sum_buffer = np.zeros(self.N, dtype=np.float32)
+        self._out_buffer = np.zeros(self.N, dtype=np.float32)
         
         # Pre-calculated reshape view (B, N) -> Project standard: [Batch, Channel]
         self._raw_bytes = np.zeros(self.payload_size, dtype=np.uint8)
@@ -38,7 +38,7 @@ class PwmDecoder:
         self._raw_bytes[:] = view
         
         # 3. Vectorized sum across ticks axis (axis=0). Written directly into preallocated buffer!
-        np.sum(self._spikes_view, axis=0, dtype=np.float16, out=self._sum_buffer)
+        np.sum(self._spikes_view, axis=0, dtype=np.float32, out=self._sum_buffer)
         
         # 4. Normalize to [0.0, 1.0] range (In-place)
         np.multiply(self._sum_buffer, self._inv_b, out=self._out_buffer)
@@ -59,12 +59,12 @@ class PopulationDecoder:
         self.payload_size = self.N * self.B
         
         # Vector of receptive field centers [0.0 ... 1.0]
-        self.centers = np.linspace(0.0, 1.0, self.M, dtype=np.float16)
+        self.centers = np.linspace(0.0, 1.0, self.M, dtype=np.float32)
         
         # Zero-Allocation Buffers
-        self._sum_buffer = np.zeros((self.V, self.M), dtype=np.float16)
-        self._mass_buffer = np.zeros(self.V, dtype=np.float16)
-        self._out_buffer = np.zeros(self.V, dtype=np.float16)
+        self._sum_buffer = np.zeros((self.V, self.M), dtype=np.float32)
+        self._mass_buffer = np.zeros(self.V, dtype=np.float32)
+        self._out_buffer = np.zeros(self.V, dtype=np.float32)
         self._silence_mask = np.zeros(self.V, dtype=bool)
 
         # [DOD FIX] Data layout is [Batch, Var, Neuron] for efficient time integration
@@ -87,7 +87,7 @@ class PopulationDecoder:
             return self._out_buffer
         
         # 3. Sum spikes across ticks (Time Integration, axis=0)
-        np.sum(self._spikes_view, axis=0, dtype=np.float16, out=self._sum_buffer) # Суммируем по тикам (ось 0)
+        np.sum(self._spikes_view, axis=0, dtype=np.float32, out=self._sum_buffer) # Суммируем по тикам (ось 0)
         
         # 4. Find total spike mass for each variable
         np.sum(self._sum_buffer, axis=1, out=self._mass_buffer)
